backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

sqlmodel>=0.0.16
sqlalchemy>=2.0

asyncpg>=0.29.0
openai==1.3.5
pydantic==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

pytest==7.4.3
pytest-asyncio==0.21.1

pydantic-settings>=2.1.0
email-validator>=2.1.0

python-dotenv>=1.0.0

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

python-multipart>=0.0.9
httpx>=0.27.0